chore: STANDPOINT in Calendar tests is now a long instead of int

+++ src/test/kotlin/CalendarTest.kt
@@ -7,13 +7,13 @@ import java.util.Calendar as JavaUtilCalendar
 
 class CalendarTest {
 
-    private val STANDPOINT = 1695492368
+    private val STANDPOINT = 1695492368L
 
     @Test
     fun `Add time to calendar`() {
         val add = 10.days
-        val standpointCalendar = Calendar.now().copy(timeInMillis = STANDPOINT.toLong())
-        val addedCalendar = Calendar.now().copy(timeInMillis = STANDPOINT.toLong() + add.inWholeMilliseconds)
+        val standpointCalendar = Calendar.now().copy(timeInMillis = STANDPOINT)
+        val addedCalendar = Calendar.now().copy(timeInMillis = STANDPOINT + add.inWholeMilliseconds)
 
         assert(standpointCalendar + add == addedCalendar) { "Addition of $add to $standpointCalendar should be $addedCalendar, but was ${standpointCalendar + add}" }
 
@@ -22,8 +22,8 @@ class CalendarTest {
     @Test
     fun `Subtract time from calendar`() {
         val subtract = 10.days
-        val standpointCalendar = Calendar.now().copy(timeInMillis = STANDPOINT.toLong())
-        val subtractedCalendar = Calendar.now().copy(timeInMillis = STANDPOINT.toLong() - subtract.inWholeMilliseconds)
+        val standpointCalendar = Calendar.now().copy(timeInMillis = STANDPOINT)
+        val subtractedCalendar = Calendar.now().copy(timeInMillis = STANDPOINT - subtract.inWholeMilliseconds)
 
         assert(standpointCalendar - subtract == subtractedCalendar) { "Subtraction of $subtract from $standpointCalendar should be $subtractedCalendar, but was ${standpointCalendar - subtract}" }
 
@@ -31,8 +31,8 @@ class CalendarTest {
 
     @Test
     fun `Comparison functionality with Java Calendar`() {
-        val calendar = Calendar.fromMilliseconds(STANDPOINT.toLong())
-        val javaCalendar = JavaUtilCalendar.getInstance().apply { timeInMillis = STANDPOINT.toLong() }
+        val calendar = Calendar.fromMilliseconds(STANDPOINT)
+        val javaCalendar = JavaUtilCalendar.getInstance().apply { timeInMillis = STANDPOINT }
 
         calendar += 20.days
         calendar += 25000.milliseconds
